调整开发环境

+++ app/src/main/java/com/guodong/utils/ImageUtils.java
@@ -551,9 +551,9 @@ public class ImageUtils {
                         pixG = Color.green(pixColor);
                         pixB = Color.blue(pixColor);
 
-                        newR = newR + (int) (pixR * gauss[idx]);
-                        newG = newG + (int) (pixG * gauss[idx]);
-                        newB = newB + (int) (pixB * gauss[idx]);
+                        newR = newR + pixR * gauss[idx];
+                        newG = newG + pixG * gauss[idx];
+                        newB = newB + pixB * gauss[idx];
                         idx++;
                     }
                 }
@@ -581,14 +581,15 @@ public class ImageUtils {
 
     /**
      * Drawable转Bitmap
+     *
      * @param drawable
      * @return
      */
-    public static Bitmap drawableToBitamp(Drawable drawable){
+    public static Bitmap drawableToBitamp(Drawable drawable) {
         int w = drawable.getIntrinsicWidth();
         int h = drawable.getIntrinsicHeight();
         Bitmap.Config config = drawable.getOpacity() != PixelFormat.OPAQUE ? Bitmap.Config.ARGB_8888 : Bitmap.Config.RGB_565;
-        Bitmap bitmap = Bitmap.createBitmap(w,h,config);
+        Bitmap bitmap = Bitmap.createBitmap(w, h, config);
         //注意，下面三行代码要用到，否在在View或者surfaceview里的canvas.drawBitmap会看不到图
         Canvas canvas = new Canvas(bitmap);
         drawable.setBounds(0, 0, w, h);
